bfg2/docs/requirements_example.txt
# BFG2 Requirements
# Add these to your project's requirements.txt

# Django Framework
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15

# Database & ORM
mysqlclient>=2.2.0
dj-database-url>=2.1.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Cache (Redis)
django-redis>=5.4.0
redis>=5.0.0
hiredis>=2.2.0

# API Documentation
drf-spectacular>=0.27.0

# Environment variables
django-environ>=0.11.2

# Image processing (for Media module)
Pillow>=10.1.0

# PDF Generation (Optional)
weasyprint>=60.0

# File storage (Optional - for S3)
boto3>=1.34.0
django-storages>=1.14.2

# Utilities
python-slugify>=8.0.1

# Development tools (Optional)
django-extensions>=3.2.3
django-debug-toolbar>=4.2.0
